backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
httpx==0.27.0
supabase==2.7.4
stripe==8.0.0
anthropic==0.49.0
slowapi==0.1.9
python-multipart==0.0.6
pydantic==2.5.3
pydantic[email]==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0
openpyxl==3.1.2
requests==2.32.5
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
aiohttp==3.9.1
keepa==1.3.8
pytrends==4.9.2
numpy==1.26.4
scipy==1.11.4
google-generativeai==0.8.3
openai==1.30.5
redis==5.0.1
bleach==6.1.0
certifi==2024.2.2
beautifulsoup4==4.12.3
lxml==5.1.0
pandas==2.2.2
spacy==3.7.4
catboost==1.2.5
apscheduler==3.10.4
plotly==5.22.0
scikit-learn==1.4.2
